modules/kvs: Rename commit.[ch] to kvstxn.[ch]
