+++ src/modules/kvs/Makefile.am
@@ -23,8 +23,8 @@ kvs_la_SOURCES = \
 	lookup.c \
 	fence.h \
 	fence.c \
-	commit.h \
-	commit.c \
+	kvstxn.h \
+	kvstxn.c \
 	msg_cb_handler.h \
 	msg_cb_handler.c \
 	kvsroot.h \
@@ -41,7 +41,7 @@ TESTS = \
 	test_cache.t \
 	test_lookup.t \
 	test_fence.t \
-	test_commit.t \
+	test_kvstxn.t \
 	test_kvs_util.t \
 	test_msg_cb_handler.t \
 	test_kvsroot.t
@@ -97,10 +97,10 @@ test_fence_t_LDADD = \
 	$(top_builddir)/src/modules/kvs/kvs_util.o \
 	$(test_ldadd)
 
-test_commit_t_SOURCES = test/commit.c
-test_commit_t_CPPFLAGS = $(test_cppflags)
-test_commit_t_LDADD = \
-	$(top_builddir)/src/modules/kvs/commit.o \
+test_kvstxn_t_SOURCES = test/kvstxn.c
+test_kvstxn_t_CPPFLAGS = $(test_cppflags)
+test_kvstxn_t_LDADD = \
+	$(top_builddir)/src/modules/kvs/kvstxn.o \
 	$(top_builddir)/src/modules/kvs/fence.o \
 	$(top_builddir)/src/modules/kvs/cache.o \
 	$(top_builddir)/src/modules/kvs/lookup.o \
@@ -126,7 +126,7 @@ test_kvsroot_t_CPPFLAGS = $(test_cppflags)
 test_kvsroot_t_LDADD = \
 	$(top_builddir)/src/modules/kvs/kvsroot.o \
 	$(top_builddir)/src/modules/kvs/waitqueue.o \
-	$(top_builddir)/src/modules/kvs/commit.o \
+	$(top_builddir)/src/modules/kvs/kvstxn.o \
 	$(top_builddir)/src/modules/kvs/cache.o \
 	$(top_builddir)/src/modules/kvs/fence.o \
 	$(top_builddir)/src/modules/kvs/msg_cb_handler.o \
